Make the engine build clean under -Wsign-compare
Fifteen comparisons tested a signed value against a size_t or an unsigned
enum, so the signed side converted to unsigned. A guard excludes a
negative value at each of them today, but the conversion is what hides
the next one.

Each site now matches its operands: a cast where the value is provably
non-negative, and the declared type where the variable is local and never
negative (check_tag's index, the changes report's bucket, htsserver's
prevlen). hts_log_vprint casts opt->debug to int rather than retyping
level, because an enum-typed level would red -Wswitch on the switch below
it.

Object code is unchanged in htscache, htsparse and punycode, and in every
htslib function but hts_log_vprint. The other three changes move a
comparison from unsigned to signed or widen a loop index, which changes
the instructions and not the result: the full suite stays green, and both
buckets tests exercise the changes report.

The six sites in src/proxy/ follow separately.

Part of #1793

+++ src/htscache.c
@@ -1310,7 +1310,7 @@ void cache_init(cache_back * cache, httrackp * opt) {
                   && zfi.size_filename < sizeof(filename) - 2) {
                 long int pos =
                   (long int) unzGetOffset((unzFile) cache->zipInput);
-                assertf(readSizeHeader < sizeof(comment));
+                assertf((size_t) readSizeHeader < sizeof(comment));
                 comment[readSizeHeader] = '\0';
                 entries++;
                 if (pos > 0) {
